requirements.txt
streamlit
pandas
openai
python-dotenv
pypdf
python-docx
vosk
soundfile
numpy
requests
tqdm
pydub
openai-whisper
torch
faster-whisper
google-generativeai
